Added OPA-python-client to requirements.txt and renamed security_group.py
