+++ requirements.txt
@@ -3,3 +3,4 @@ hydra-core>=1.3.2
 azure-keyvault-secrets>=4.8.0
 azure-identity>=1.15.0
 PyGithub>=2.2.0
+OPA-python-client>=1.3.2
